Create `Subject` (close #13)
This adds Subject to the tracker, along with updating docs and tests.

Subject allows you to attach a subject to either a tracker to send with all events, and/or attach a subject to an event, which will take precedence over the tracker-level subject

Payload Tidying:

The structured event fields in Payload have been replaced with a single instance of Option<StructuredEvent>. This allows us to move the serialization handling to the existing StructuredEvent struct by applying the #[serde(flatten)] attribute to the structured_event field of Payload, which (as expected) flattens the struct to inline the fields into Payload on serialization.