Ensure no panics (close #15)
This removes any runtime panics that could happen within the library. This involved creating the `Error` Enum as a return type for possible failures, and changing our return types to `Result<T, Error>`
Docs have also been updated to reflect new expected usage of functions that return Result.

There are currently two errors in the Enum:

- BuilderError
This, using the #[builder(build_fn(error = "Error"))] attribute on builder structs, handles errors that occur at any point during the building events or payloads.

- EmitterError
This handles any errors that can occur in the emitter. It currently only exists in Emitter.post, but will likely be expanded as the emitter is built out.