Add Event Store (close #24)
This adds an Event Store to the tracker and several other changes required to facilitate the event store's function and the tracker's future extensibility.

= What's New =

- The async-trait crate has been added, to allow for the HttpClient.post trait method to be async. Currently, async traits are unstable, so this crate is a requirement until the feature is available in stable rust.
- `Emitter`, `EventStore`, and `HttpClient` traits added to allow users to create their own implementations of each
- `InMemoryEventStore` added, an implementation of the `EventStore` trait, which queues events in a Vec
- The `EventBatch` struct was added to contain the batched events returned from the event store.
- EventStoreError was added to the Error enum.

= What's Changed =

== BatchEmitter (previously called Emitter) ==

The previous Emitter (now named BatchEmitter) has been updated to utilise an implementation of the EventStore trait, in this case, InMemoryEventStore.

The BatchEmitter is an implementation of the new Emitter trait. The batch emitter is split into two distinct parts.
1. The main Sync program
2. The Async Tokio runtime loop running in its own thread

This pattern was chosen as it mean we aren't required to make all previous functions async to accommodate Reqwest::post being an async call, whilst still being able to utilise tokio to handle the spawning of any number of async tasks from the main sync program. It also prevents any users from having to handle the async side themselves if they aren't using the tracker on an async setup.

Communication between sync and async is done using tokio channels.

== PayloadBuildable ==
This trait has been re-named to EventAddable, as we no longer want to finish building the event immediately after adding the event to the payload, due to needing to set dvce_sent_tstamp in the event store.

The associated method has also been renamed from build_payload to add_to_payload

== Tracker ==

- Tracker.track is no longer an async method, removing the forced requirement for the user program to implement async
- Tracker.emitter now accepts any implementation of Emitter
- The emitter_mut method has been added. This was necessary as closing down the emitter requires a mutable reference, as it mutates data within the event store, owned by the emitter.

== Payload ==
The finalise_payload method was added to set dvce_sent_tstamp when the event is taken out the event store, rather than within the Tracker.track method

== Tests ==
- Common functionality between tests has been moved to /common
- Due to a bug in Micro, it is required to pass an Iglu Resolver config file to prevent responses from taking a long time. This required implementing the ImageArgs and Image traits for Micro and can be found in ./tests/common/micro.rs.
- Add integration tests for BatchEmitter