Add Retry (close #30)
This adds a retry mechanism to the `BatchEmitter`, along with a few other adjustments to make this possible.

= Crates =

`rand` crate has been added for creating "jitter" in the retry delay.

= RetryPolicy =

The `RetryPolicy` enum has been added to manage how events are handled in the `BatchEmitter` when they don't successfully send. It has three options for a failed event:
- Never retry
- Retry n times
- Retry forever

= EventBatch =

`id` and `delay` and `retry_attempts` fields have been added. The event batch `id` is a `UUID`, taken from the first event in the vec of events in an `EventBatch` and is used to identify the batch in logging. The `delay` field is used to store the duration to delay sending by in retry. `retry_attempts` keeps track of how many times an event has attempted to send, used in conjunction with a `RetryPolicy`.

- add `update_event_stm` and `update_batch_delay` methods to update the sent time for events in a batch and handle batch delay logic respectively.

= BatchEmitter =

== Builder ==

The batch emitter builder now accepts a generic `HttpClient` implementation to the `http_client` field. This was done to be able to pass in the `FlakyHttpClient` in testing, but is equally useful for users to be able to create their own `HttpClient` implementations.

If no `HttpClient` is provided, `ReqwestClient` is used as a default.

== Retry ==

The retry mechanism for the batch emitter uses another tokio channel to attempt to re-send failed events batches. This channel (with it's components shown as `RetryReceiver` and `RetryTransmitter` in the diagram below) acts as both a store for retrying events and the method of retrying. It is an unbounded channel (e.g. the number of messages it can contain is not restricted), to match capacity of the "sending" HashMaps used in other trackers.

We make use of the existing `EventBatch` by assigning/updating its delay, and updating `stm` of each contained event before another send attempt.

An `EventBatch` delay is exponentially increased to a maximum of 10 minutes, with between 0-1s of jitter added/subtracted from a delay multiplier of 2, effectively giving us a delay increase of `prev_delay_time * random_number_between(1, 3)`.

== send_batch ==

The return type of `send_batch` has been changed from `Result<(), Error>` to `Result<SentEventResponse, EventBatch>`

`send_batch` now returns the passed in `EventBatch` if the batch fails to send. This allows us not to have to perform a `clone` on the EventBatch for every batch (which may be rather large). The clone was required as, previously, the event batch was moved into `send_batch`. The batch id is returned on success for logging.

== Flush ==

Flush now works by creating and sending `full_batch`s until there aren't enough to create a full batch, at which point a batch is created of any remaining events in the event store, rather than sending one huge batch of every event in the store.

== Other minor BatchEmitter changes ==

- batch id has been included in logging where possible
- the debug log in `rx.recv()` has been removed, as the output of that log included the entire string of every event in the batch
- removed some redundant borrow/mut for `rx`
- removed un-needed cloning of `tx`
- adjust "waiting for task" debug log to start from 1, rather than 0

= Payload =

- `stm` is now `pub (crate)` to allow us to access and modify when retrying

= Test =

- add `FlakyHttpClient` to test retry, which can fail to send `n` times until sending successfully
- added a timeout to `wait_for_events` to prevent tests running infinitely in the case of events never being present
- updated `std::time::sleep` to `tokio::time::sleep` to allow any other async tasks to continue